Termination handler stops multibar to prevent cursor issues
